requirements.txt
flask
flask-sqlalchemy
flask-wtf
psycopg2
python-slugify
gunicorn
flask_restful